Add FritzDECT charger
